Add HTTP status code and hanle exceptions

+++ src/main/java/com/dauphine/blogger/services/implementations/CategoryServiceImplementation.java
@@ -1,5 +1,7 @@
 package com.dauphine.blogger.services.implementations;
 
+import com.dauphine.blogger.exceptions.CategoryAlreadyExistsException;
+import com.dauphine.blogger.exceptions.CategoryNotFoundByIdException;
 import com.dauphine.blogger.models.Category;
 import com.dauphine.blogger.repositories.CategoryRepository;
 import com.dauphine.blogger.services.CategoryService;
@@ -9,22 +11,35 @@ import java.util.List;
 import java.util.UUID;
 
 /**
+ * <p>
  * Implementation of the CategoryService interface.
- * Provides methods to manage categories, including retrieving, creating, updating,
- * and deleting categories.
+ * Provides methods to manage categories, including retrieving, creating, updating, and deleting categories.
+ * </p>
  *
- * @author Nelson PROIA <nelson.proia@dauphine.eu>
+ * <p>
+ * All methods in this class may throw {@link CategoryNotFoundByIdException} if the specified category ID does not exist.
+ * For methods that involve creating or updating categories, they may also throw {@link CategoryAlreadyExistsException} if a category with the same name already exists.
+ * </p>
+ *
+ * <p>
+ * This class is responsible for implementing the business logic associated with category management.
+ * It interacts with the underlying data source through CategoryRepository.
+ * </p>
+ *
+ * <p>
+ * Author: Nelson PROIA <nelson.proia@dauphine.eu>
+ * </p>
  */
 @Service
 public class CategoryServiceImplementation implements CategoryService {
 
     /**
-     * Repository for managing categories.
+     * Repository for managing category entities.
      */
     private final CategoryRepository categoryRepository;
 
     /**
-     * Constructs a CategoryServiceImplementation object with the specified CategoryRepository.
+     * Constructs a new CategoryServiceImplementation object with the specified CategoryRepository.
      *
      * @param categoryRepository The repository for managing categories
      */
@@ -43,16 +58,27 @@ public class CategoryServiceImplementation implements CategoryService {
     }
 
     /**
+     * Retrieves categories by name.
+     *
+     * @param name The name of the categories to retrieve
+     * @return A list of categories with the specified name
+     */
+    @Override
+    public List<Category> getCategoriesByName(String name) {
+        return categoryRepository.findAllByName(name);
+    }
+
+    /**
      * Retrieves a category by its ID.
      *
      * @param id The ID of the category to retrieve
-     * @return The category with the specified ID, or null if not found
-     * @throws RuntimeException if the category with the specified ID does not exist
+     * @return The category with the specified ID
+     * @throws CategoryNotFoundByIdException if the category with the specified ID does not exist
      */
     @Override
-    public Category getCategory(UUID id) {
+    public Category getCategory(UUID id) throws CategoryNotFoundByIdException {
         return categoryRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Category with id " + id + " does not exist!"));
+                .orElseThrow(() -> new CategoryNotFoundByIdException(id));
     }
 
     /**
@@ -60,14 +86,14 @@ public class CategoryServiceImplementation implements CategoryService {
      *
      * @param name The name of the new category
      * @return The newly created category
-     * @throws RuntimeException if a category with the same name already exists
+     * @throws CategoryAlreadyExistsException if a category with the same name already exists
      */
     @Override
-    public Category createCategory(String name) {
+    public Category createCategory(String name) throws CategoryAlreadyExistsException {
         final boolean alreadyExistsByName = categoryRepository.existsByName(name);
 
         if (alreadyExistsByName) {
-            throw new RuntimeException("Category with name " + name + " already exists!");
+            throw new CategoryAlreadyExistsException(name);
         }
 
         Category category = new Category(name);
@@ -80,11 +106,19 @@ public class CategoryServiceImplementation implements CategoryService {
      *
      * @param id   The ID of the category to update
      * @param name The new name for the category
-     * @return The updated category, or null if the category with the specified ID was not found
+     * @return The updated category
+     * @throws CategoryNotFoundByIdException  if the category with the specified ID does not exist
+     * @throws CategoryAlreadyExistsException if a category with the same name already exists
      */
     @Override
-    public Category updateCategoryName(UUID id, String name) {
-        Category category = getCategory(id);
+    public Category updateCategoryName(UUID id, String name) throws CategoryNotFoundByIdException, CategoryAlreadyExistsException {
+        final Category category = getCategory(id);
+
+        final boolean alreadyExistsByName = categoryRepository.existsByName(name);
+
+        if (alreadyExistsByName) {
+            throw new CategoryAlreadyExistsException(name);
+        }
 
         category.setName(name);
 
@@ -96,14 +130,19 @@ public class CategoryServiceImplementation implements CategoryService {
      *
      * @param id The ID of the category to delete
      * @return true if the category was deleted successfully, false otherwise
+     * @throws CategoryNotFoundByIdException if the category with the specified ID does not exist
      */
     @Override
-    public boolean deleteCategory(UUID id) {
-        boolean deleted = categoryRepository.existsById(id);
+    public boolean deleteCategory(UUID id) throws CategoryNotFoundByIdException {
+        final boolean exists = categoryRepository.existsById(id);
+
+        if (!exists) {
+            throw new CategoryNotFoundByIdException(id);
+        }
 
         categoryRepository.deleteById(id);
 
-        return deleted;
+        return true;
     }
 
 }
